METAMAC-1457 Order de tipo Integer en vez de Long
git-svn-id: http://svn.arte-consultores.com/svn/arte/istac/metamac/metamac-srm/trunk/metamac-srm@44101 05f23f4b-1f0b-0410-ba1b-ad9509270e59

+++ metamac-srm-web/src/main/java/org/siemac/metamac/srm/web/code/presenter/CodePresenter.java
@@ -241,7 +241,7 @@ public class CodePresenter extends Presenter<CodePresenter.CodeView, CodePresent
     }
 
     @Override
-    public void updateCodeInOrder(String codeUrn, final String codelistOrderIdentifier, Long newCodeIndex) {
+    public void updateCodeInOrder(String codeUrn, final String codelistOrderIdentifier, Integer newCodeIndex) {
         throw new UnsupportedOperationException();
     }
